shortenner/requirements.txt
djangorestframework
markdown
django-filter
django-environ
psycopg2-binary
gunicorn
django_replicated
django-prometheus